// csp-app/src/app/models/data-table-structure-model.ts
export class DataTableStructureModel {
    id: number;
    reporT_ID: number;
    fielD_SORT: number;
    fielD_NAME: string;
    fielD_DISPLAY_NAME: string;
    charT_FIELD_NAME: string;
    datA_TYPE: string;
    requireD_FIELD: boolean;
    dB_INCLUDE: boolean;
    createD_BY: string
    createD_DATE: Date;
    updateD_BY: string;
    updateD_DATE: Date;
    isavtive: boolean;
}